Update 2023.08.11.15
To test the rendering created an Object based from the console log results of the useEffect. Then created a container classed div below the search classed div that renders a card  with a movie classed div. The card would show the movie's year, poster, type, and title. There is an tetiary if statement if no poster is available to default to an online placeholder source